+++ app/helpers.php
@@ -1,0 +1,10 @@
+<?php
+if (! function_exists('mb_ucfirst')) {
+    function mb_ucfirst($string, $encoding = 'UTF-8')
+    {
+        $strlen = mb_strlen($string, $encoding);
+        $firstChar = mb_substr($string, 0, 1, $encoding);
+        $then = mb_substr($string, 1, $strlen - 1, $encoding);
+        return mb_strtoupper($firstChar, $encoding) . $then;
+    }
+}