+++ gulpfile.js
@@ -128,7 +128,21 @@ gulp.task('ckeditor', function () {
         .pipe(gulp.dest('public/components/ckeditor/lang'));
 
     // Plugins
-    var plugins = ['clipboard', 'image', 'image2', 'justify', 'lineutils', 'link', 'magicline', 'panelbutton', 'showblocks', 'specialchar', 'table', 'widget'];
+    var plugins = [
+        'clipboard',
+        'image',
+        'image2',
+        'justify',
+        'lineutils',
+        'link',
+        'magicline',
+        'panelbutton',
+        'pastefromword',
+        'showblocks',
+        'specialchar',
+        'table',
+        'widget'
+    ];
     for (var i = 0; i < plugins.length; i++) {
         gulp.src(['node_modules/ckeditor/plugins/' + plugins[i] + '/**/*'])
             .pipe(gulp.dest('public/components/ckeditor/plugins/' + plugins[i]));
@@ -146,9 +160,8 @@ gulp.task('js-admin', function () {
             'node_modules/angular-resource/angular-resource.js',
             'node_modules/angular-smart-table/dist/smart-table.js',
             'node_modules/angular-ui-tree/dist/angular-ui-tree.js',
-            'node_modules/vue/vue.js',
-            'node_modules/vue-tables/lib/dist/v-client-table.js',
-            'node_modules/jsurl/url.js',
+            // 'node_modules/vue/vue.js',
+            // 'node_modules/vue-tables/lib/dist/v-client-table.js',
             'node_modules/bootstrap/js/dropdown.js',
             'node_modules/bootstrap/js/collapse.js',
             'node_modules/bootstrap/js/alert.js',
@@ -164,8 +177,8 @@ gulp.task('js-admin', function () {
 
     return gulp.src(files)
         .pipe(concat('components.js'))
-        // .pipe(ngAnnotate())
-        // .pipe(uglify())
+        .pipe(ngAnnotate())
+        .pipe(uglify())
         .on('error', swallowError)
         .pipe(rename(destFile))
         .pipe(gulp.dest(destDir));