+++ resources/assets/js/public/offcanvas.js
@@ -1,0 +1,6 @@
+$(function () {
+    $('[data-toggle="offcanvas"]').click(function (e) {
+        $('.sidebar-offcanvas').toggleClass('active');
+        e.preventDefault();
+    });
+});