+++ resources/lang/es/validation.php
@@ -192,6 +192,7 @@ return [
         'DDMMYYYY'      => 'DD.MM.YYYY',
         'DDMMYYYY HHMM' => 'DD.MM.YYYY HH:MM',
         'location'      => 'Ubicación',
+        'venue'         => 'Ubicación',
         'price'         => 'Precio',
         'currency'      => 'Moneda',